commit drivers and shifts initialise from excel
